refactor: move all source code into `src` folder, except for file in `app` dir
